+++ src/index.ts
@@ -2,7 +2,21 @@ import { Client } from './classes/Client';
 
 import { Status } from './enums/Status';
 
+import { AuthPersonalCredentials } from './interfaces/auth/PersonalCredentials';
+import { Host } from './interfaces/host/Host';
+import { GetHostsQueryParams } from './interfaces/host/GetQueryParams';
+import { GetHostsPayload } from './interfaces/host/GetPayload';
+
 import { InvalidCredentialsError } from './errors/InvalidCredentials';
 import { TFARequiredError } from './errors/TFARequired';
 
-export { Client, Status, InvalidCredentialsError, TFARequiredError };
+export {
+  Client,
+  Status,
+  AuthPersonalCredentials,
+  Host,
+  GetHostsQueryParams,
+  GetHostsPayload,
+  InvalidCredentialsError,
+  TFARequiredError
+};