+++ src/interfaces/host/GetPayload.ts
@@ -1,0 +1,7 @@
+import { Host } from './Host';
+
+/** Payload of the `getHosts` call */
+export interface GetHostsPayload {
+  data: Host[];
+  has_more: boolean;
+}