+++ src/interfaces/host/GetQueryParams.ts
@@ -1,0 +1,10 @@
+/**
+ * Object representation of `getHosts` request query params
+ */
+export interface GetHostsQueryParams {
+  /** Host's mode */
+  mode: 'desktop' | 'game';
+
+  /** Host's visibility */
+  public?: boolean;
+}